+++ lib/circular_text.dart
@@ -1,3 +1,4 @@
 library circular_text;
 
+export 'circular_text/model.dart';
 export 'circular_text/widget.dart';