+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.0
+* Added multi `Text` support
+* Added `StartAngleAlignment` property to align text around `startAngle`
+* Added single and multi text examples
+
 ## 0.2.1
 * Added `Text`, changed angle position and added some minor changes